requirements.txt
# Django core
Django==5.1.2
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-celery-results==2.5.1
python-decouple==3.8
psycopg2-binary==2.9.10

# Celery + Redis
celery==5.4.0
redis==5.2.0
flower==2.0.1


# Machine Learning / Predicción
tensorflow==2.19.0  # Usas 2.19 en tu viejo req
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.1.3

# Exportación PDF
weasyprint==65.1
pypdf==5.4.0
reportlab==4.2.5
# xhtml2pdf==0.2.15 # si usas este, sino omitir

# Excel
openpyxl==3.1.5

# Web Scraping y utilidades
beautifulsoup4==4.13.4
requests==2.32.3

# Otros que vi en tu código
matplotlib==3.10.1
pydot==3.0.4
graphviz==0.20.3

# Gunicorn para prod server
gunicorn==21.2.0
